Implelemt game() function
The game function plays 5 rounds. It prompts the use to input either
rock, paper or scissors. Then the use input is validated whether
they in fact chose either of these three. If the player didn't make
a valid choice the round is played again. I it's correct, the game is played
as expected. Every round a score is kept and at the end of these 5 rounds
a winner is determined.